+++ test/standalone/test_louvain.cpp
@@ -18,7 +18,7 @@ TEST(Louvain, DenseTest)
     int num_nodes = test_adj_mat.size();
     vector<unsigned long> degrees;
     vector<unsigned int> links;
-    vector<float> weights;
+    vector<half> weights;
 
     int cum_degree = 0;
     for (int i = 0; i < num_nodes; i++)
@@ -29,7 +29,7 @@ TEST(Louvain, DenseTest)
             {
                 cum_degree += 1;
                 links.push_back(j);
-                weights.push_back(test_adj_mat[i][j]);
+                weights.push_back((half)test_adj_mat[i][j]);
             }
         }
         degrees.push_back(cum_degree);
@@ -39,55 +39,55 @@ TEST(Louvain, DenseTest)
 
     ASSERT_EQ(g.get_num_neighbors(0), 2);
     ASSERT_EQ(g.get_num_neighbors(3), 1);
-    ASSERT_EQ(g.get_num_selfloops(0), 0);
-    ASSERT_EQ(g.get_weighted_degree(0), 2);
-    ASSERT_EQ(g.get_weighted_degree(2), 2);
-    ASSERT_EQ(g.total_weight, 14);
+    ASSERT_EQ(g.get_num_selfloops(0), (half)0);
+    ASSERT_EQ(g.get_weighted_degree(0), (half)2);
+    ASSERT_EQ(g.get_weighted_degree(2), (half)2);
+    ASSERT_EQ(g.total_weight, (half)14);
 
-    Community c = Community(g, -1, 0.000001);
-    ASSERT_NEAR(c.modularity(), -0.1326530612244898, 1e-6);
+    Community c = Community(g, -1, (half)0.000001);
+    ASSERT_NEAR(c.modularity(), (half)-0.1326530612244898, 1e-6);
     c.compute_neigh_comms(0);
     ASSERT_EQ(c.neigh_last, 3);
 
     vector<float> test_neigh_weight = {0, -1, 1, -1, -1, -1, -1, 1};
     for (int i = 0; i < c.num_nodes; i++)
     {
-        ASSERT_EQ(c.neigh_weight[i], test_neigh_weight[i]);
+        ASSERT_EQ(c.neigh_weight[i], (half)test_neigh_weight[i]);
     }
 }
 
 TEST(Louvain, SparseTest)
 {
-    vector<vector<float>> test_adj_mat = {{0, 0, 1, 0, 0, 0, 0, 1},
-                                          {0, 0, 0, 0, 1, 1, 0, 0},
-                                          {1, 0, 0, 0, 0, 0, 0, 1},
-                                          {0, 0, 0, 0, 0, 0, 1, 0},
-                                          {0, 1, 0, 0, 0, 1, 0, 0},
-                                          {0, 1, 0, 0, 1, 0, 0, 0},
-                                          {0, 0, 0, 1, 0, 0, 0, 0},
-                                          {1, 0, 1, 0, 0, 0, 0, 0}};
+    vector<vector<half>> test_adj_mat = {{(half)0, (half)0, (half)1, (half)0, (half)0, (half)0, (half)0, (half)1},
+                                         {(half)0, (half)0, (half)0, (half)0, (half)1, (half)1, (half)0, (half)0},
+                                         {(half)1, (half)0, (half)0, (half)0, (half)0, (half)0, (half)0, (half)1},
+                                         {(half)0, (half)0, (half)0, (half)0, (half)0, (half)0, (half)1, (half)0},
+                                         {(half)0, (half)1, (half)0, (half)0, (half)0, (half)1, (half)0, (half)0},
+                                         {(half)0, (half)1, (half)0, (half)0, (half)1, (half)0, (half)0, (half)0},
+                                         {(half)0, (half)0, (half)0, (half)1, (half)0, (half)0, (half)0, (half)0},
+                                         {(half)1, (half)0, (half)1, (half)0, (half)0, (half)0, (half)0, (half)0}};
 
     int num_nodes = test_adj_mat.size();
-    SparseMatrixDOK<float> sm = SparseMatrixDOK<float>(num_nodes, num_nodes, 0, false, true);
+    SparseMatrixDOK<half> sm = SparseMatrixDOK<half>(num_nodes, num_nodes, (half)0, false, true);
     sm.from_densematrix(test_adj_mat);
 
     Graph g = Graph(sm);
 
     ASSERT_EQ(g.get_num_neighbors(0), 2);
     ASSERT_EQ(g.get_num_neighbors(3), 1);
-    ASSERT_EQ(g.get_num_selfloops(0), 0);
-    ASSERT_EQ(g.get_weighted_degree(0), 2);
-    ASSERT_EQ(g.get_weighted_degree(2), 2);
-    ASSERT_EQ(g.total_weight, 14);
+    ASSERT_EQ(g.get_num_selfloops(0), (half)0);
+    ASSERT_EQ(g.get_weighted_degree(0), (half)2);
+    ASSERT_EQ(g.get_weighted_degree(2), (half)2);
+    ASSERT_EQ(g.total_weight, (half)14);
 
-    Community c = Community(g, -1, 0.000001);
-    ASSERT_NEAR(c.modularity(), -0.1326530612244898, 1e-6);
+    Community c = Community(g, -1, (half)0.000001);
+    ASSERT_NEAR(c.modularity(), (half)-0.1326530612244898, 1e-6);
     c.compute_neigh_comms(0);
     ASSERT_EQ(c.neigh_last, 3);
 
     vector<float> test_neigh_weight = {0, -1, 1, -1, -1, -1, -1, 1};
     for (int i = 0; i < c.num_nodes; i++)
     {
-        ASSERT_EQ(c.neigh_weight[i], test_neigh_weight[i]);
+        ASSERT_EQ(c.neigh_weight[i], (half)test_neigh_weight[i]);
     }
 }
